requirements.txt
psycopg2-binary==2.9.5
beautifulsoup4==4.11.1
requests==2.28.2
wikipedia==1.4.0
networkx==3.0
pandas==1.5.2
pre-commit==2.21.0